contractors/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
passlib==1.7.4
bcrypt==4.0.1
tortoise-orm[asyncmy]
python-dotenv == 1.0.0